refactor : 엔티티 및 dto 컨벤션에 의한 리팩터링
