+++ src/main/java/com/example/demo/domain/board/domain/dto/page/PageSort.java
@@ -1,0 +1,6 @@
+package com.example.demo.domain.board.domain.dto.page;
+
+public enum PageSort {
+    ASC,
+    DESC
+}
