--- src/main/java/com/example/demo/domain/board/domain/page/PageSort.java
@@ -1,6 +1,0 @@
-package com.example.demo.domain.board.domain.page;
-
-public enum PageSort {
-    ASC,
-    DESC
-}
